+++ Parsing/Request/Request.cpp
@@ -185,7 +185,7 @@ void Request::parseHeaderLines(Config config)
 
     size_t pos = 0;
     // std::cout << "i am here" << std::endl;
-    for (int i = 1; i < headerPart.size(); i++)
+    for (size_t i = 1; i < headerPart.size(); i++)
     {
         if ((pos = headerPart[i].find(":")) != std::string::npos)
             headerMap[headerPart[i].substr(0, pos)] = headerPart[i].substr(pos + 2, headerPart[i].find("\r\n") - pos - 2);   
@@ -215,7 +215,7 @@ int Request::parse_body(std::string c)
 
  void Request::print_header()
  {
-     for (int i = 0; i < headerPart.size(); i++)
+     for (size_t i = 0; i < headerPart.size(); i++)
         std::cout <<  "|" << headerPart[i] <<  "|" << std::endl;
     std::cout << save << std::endl;
  }
@@ -273,7 +273,7 @@ void Request::parseUri()
     int Request::checkURI()
     {
         std::string allowedchars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~:/?#[]@!$&'()*+,;=%";
-        for (int i = 0; i < uri.size(); i++)
+        for (size_t i = 0; i < uri.size(); i++)
         {
             if (allowedchars.find(uri[i]) == std::string::npos)
                 throw "409";
@@ -295,7 +295,7 @@ void Request::parseUri()
 
     void Request::checkContentLength(int a)
     {
-        if (contentLength > location.getClientMaxBodySize()
+        if (contentLength > (long long)location.getClientMaxBodySize()
             || contentLength < 0)
         {
             throw "413";
